+++ src/kptncook/__init__.py
@@ -44,7 +44,7 @@ __all__ = [
     "export_recipes_to_tandoor",
 ]
 
-__version__ = "0.0.28"
+__version__ = "0.0.29"
 logger = logging.getLogger(__name__)
 cli = typer.Typer()